// types/types.ts
export interface Invoice {
  id: number;
  staff: InvoiceCustomer;
  customer: InvoiceCustomer;
  invoice_items: InvoiceItem[];
  invoice_number: string;
  status: string;
  total_price: string;
  discounts: string;
  shipping_costs: string;
  taxes: string;
  amount_due: string;
  payment_terms: string;
  additional_notes: string;
  timestamp: Date;
  updated: Date;
}

export interface InvoiceCustomer {
  id: number;
  first_name: string;
  last_name: string;
  email: string;
  phone_number: string | null;
  profile_pic: string | null;
}

export interface InvoiceItem {
  id: number;
  product: InvoiceItemProduct;
  quantity: number;
  discounts: number;
}

export interface InvoiceItemProduct {
  id: number;
  name: string;
  price: number;
}


export interface Category {
  id?: number,
  title: string,
  description: string,
}


/* KYC */
export interface Kyc {
  id: number;
  user_type: string;
  email: string;
  first_name: string;
  last_name: string;
  phone_number: string;
  dob: string;
  id_type: string;
  id_number: string;
  country: string;
  location: string;
  id_front: string | null;
  id_back: string | null;
  profession: string;
  father_name: string;
  mother_name: string;
  witness_name: string;
  witness_relation: string;
  user_age: string;
  gender: string;
  marital_status: string;
  religion: string;
  medication: boolean;
  medication_type: string;
  childrens: number;
  boys: number;
  girls: number;
  banks: string;
  doc: string | null;
  date_joined: string;
  profile_pic: string | null
}

/* COPARATE */
export interface Coparate {
  id: number;
  name: string;
  courier: string;
  email: string;
  province: string;
  exact_location: string;
  mobile_number: string;
  created_at: Date;
  updated_at: Date;
}


/* PRODUCT */
export interface Product {
  id: number;
  thumb: string;
  name: string;
  description: string;
  color: null;
  size: null;
  img: null | string;
  author: null | Author
  price: string;
  discount: string;
  is_pub: boolean;
  category: ProductCategory;
}

export interface ProductCategory {
  id: number;
  cover: string;
  title: string;
  timestamp: Date;
  updated: Date;
}

export interface ProductDefaultParams {
  id?: number | null;
  name: string;
  description: string;
  price: number | string;
}

export interface Author {
  id: number;
  first_name: string;
  last_name: string;
  phone_number: null;
  email: string;
  location: string;
}